# swagger.yaml
openapi: 3.0.0

info:
  title: cLOUDy UBL Ordering
  version: 1.0.0
  description: |
    This is the UBL ordering API.

    🔐 Authentication  
    All endpoints except `/create-key`, `/health`, and `/v1/validate-xml` require an API key.

    To obtain an API key:
    1. Call **POST /get-key**
    2. Copy the returned key
    3. Send it in requests:

    api-key: YOUR_API_KEY

servers:
  - url: http://cloudylb-504707632.us-east-1.elb.amazonaws.com/api

security:
  - ApiKeyAuth: []

tags:
  - name: Auth
    description: Operations for authentication
  - name: Orders
    description: Operations about the orders
  - name: Validation
    description: Operations for XML validation
  - name: System
    description: System health and monitoring

paths:

  /health:
    get:
      tags:
        - System
      security: []
      summary: Health check
      description: Returns the health status of the API service.
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "running"

        "500":
          description: Service is unhealthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'

  /get-key:
    post:
      tags:
        - Auth
      security: []
      summary: Generate or get API key
      description: Authenticate a user and generate an API key. If user already registered returns their key.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'

      responses:
        "200":
          description: API key generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponse'

        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'

        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'

  /v1/buyer/{buyerId}/order:
    post:
      tags:
        - Orders
      summary: Create order
      description: Creates an order for a buyer and returns the generated UBL XML.
      parameters:
        - name: buyerId
          in: path
          required: true
          schema:
            type: string
            format: uuid

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'

      responses:
        "200":
          description: Order created successfully
          content:
            application/xml:
              schema:
                type: string
                example: "<Order>...</Order>"

        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'

        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidApiKeyError'

        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'

    get:
      tags:
        - Orders
      summary: List orders for a buyer
      description: Retrieve a paginated list of orders for a buyer.
      parameters:
        - name: buyerId
          in: path
          required: true
          schema:
            type: string
            format: uuid

        - name: status
          in: query
          required: false
          schema:
            type: string
            example: CREATED

        - name: fromDate
          in: query
          required: false
          schema:
            type: string
            format: date
            example: "2026-03-01"

        - name: toDate
          in: query
          required: false
          schema:
            type: string
            format: date
            example: "2026-03-10"

        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10

        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0

      responses:
        "200":
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdersListResponse'

        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'

        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidApiKeyError'

        "404":
          description: Buyer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'

        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'

  /v1/buyer/{buyerId}/order/{orderId}:
    get:
      tags:
        - Orders
      summary: Get order by ID
      description: Retrieve a specific order for a buyer.
      parameters:
        - name: buyerId
          in: path
          required: true
          schema:
            type: string
            format: uuid

        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid

      responses:
        "200":
          description: Order retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'

        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidApiKeyError'

        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'

        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'

    put:
      tags:
        - Orders
      summary: Update order
      description: Update an existing order for a buyer and return the updated UBL XML document.
      parameters:
        - name: buyerId
          in: path
          required: true
          schema:
            type: string
            format: uuid

        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdateRequest'

      responses:
        "200":
          description: Order updated successfully
          content:
            application/xml:
              schema:
                type: string
                example: "<Order>...</Order>"

        "400":
          description: Invalid request or validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'

        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidApiKeyError'

        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'

        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'

    delete:
      tags:
        - Orders
      summary: Delete order
      description: Delete an existing order for the given buyer and orderId, only if the order status is already CANCELED.
      parameters:
        - name: buyerId
          in: path
          required: true
          schema:
            type: string
            format: uuid

        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid

      responses:
        "200":
          description: Order deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteOrderResponse'

        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidApiKeyError'

        "403":
          description: Buyer does not have access to this order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'

        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'

        "409":
          description: Order cannot be deleted unless status is CANCELED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'

        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'

  /v1/buyer/{buyerId}/order/CANCELED:
    delete:
      tags:
        - Orders
      summary: Delete all cancelled orders for buyer
      description: Delete all orders with status CANCELED belonging to the given buyerId.
      parameters:
        - name: buyerId
          in: path
          required: true
          schema:
            type: string
            format: uuid

      responses:
        "200":
          description: All cancelled orders deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAllCancelledOrdersResponse'

        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidApiKeyError'

        "404":
          description: Buyer not found or buyerId invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'

        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'

  /v1/validate-xml:
    post:
      tags:
        - Validation
      security: []
      summary: Validate XML
      description: Validate an order XML document.
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              type: string
              example: "<Order>...</Order>"

      responses:
        "200":
          description: XML validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/XmlValidationResponse'

        "400":
          description: XML validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/XmlValidationError'

        "500":
          description: Server error during validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: api-key

  schemas:
    CreateApiKeyRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "john_doe"
        password:
          type: string
          example: "password123"

    ApiKeyResponse:
      type: object
      properties:
        apikey:
          type: string
          example: "ubl_sk_b7c4c5e8a9d34f6c9b8e21d7a4f12345"

    OrderRequest:
      type: object
      required:
        - order_date
        - delivery_date
        - currency_code
        - address
        - items
      properties:
        order_date:
          type: string
          format: date
          example: "2026-02-27"
        delivery_date:
          type: string
          format: date
          example: "2026-03-10"
        currency_code:
          type: string
          example: "AUD"
        address:
          $ref: '#/components/schemas/Address'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'

    OrderUpdateRequest:
      type: object
      properties:
        order_date:
          type: string
          format: date
          example: "2026-02-27"
        delivery_date:
          type: string
          format: date
          example: "2026-03-10"
        currency_code:
          type: string
          example: "AUD"
        address:
          $ref: '#/components/schemas/Address'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'

    Address:
      type: object
      properties:
        street:
          type: string
          example: "436 George St"
        city:
          type: string
          example: "Sydney"
        state:
          type: string
          example: "NSW"
        postal_code:
          type: string
          example: "2000"
        country_code:
          type: string
          example: "AU"

    OrderItem:
      type: object
      properties:
        item_name:
          type: string
          example: "Widget"
        item_description:
          type: string
          example: "Industrial widget"
        quantity:
          type: integer
          example: 15
        unit_price:
          type: number
          example: 50

    OrderResponse:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
          example: "1273819kadshj1231j123kasdkja"
        productId:
          type: string
          nullable: true
          example: "a9c3d2f1-1234-4f6a-b8e2-11d7a4f12345"
        quantity:
          type: integer
          example: 10
        money:
          type: number
          example: 500
        status:
          type: string
          example: "CREATED"
        xml:
          type: string
          example: "<Order>...</Order>"

    OrdersListResponse:
      type: object
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: "All orders for a buyer"
        buyerId:
          type: string
          format: uuid
          example: "0d2f9e2a-1f40-4b12-9e22-1b2a9e0c1234"
        count:
          type: integer
          example: 2
        limit:
          type: integer
          example: 10
        offset:
          type: integer
          example: 0
        orders:
          type: array
          items:
            type: object
            properties:
              orderId:
                type: string
                format: uuid
              status:
                type: string
                example: "CREATED"
              orderDate:
                type: string
                format: date
                example: "2026-03-10"
              deliveryDate:
                type: string
                format: date
                example: "2026-03-12"
              currencyCode:
                type: string
                example: "AUD"
              itemCount:
                type: integer
                example: 1
              totalAmount:
                type: string
                example: "100"

    DeleteOrderResponse:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
          example: "1273819kadshj1231j123kasdkja"
        message:
          type: string
          example: "Order deleted successfully"

    DeleteAllCancelledOrdersResponse:
      type: object
      properties:
        buyerId:
          type: string
          format: uuid
          example: "2c0c7e5b-1f1e-4a5e-90ef-8a3d3a0a3d11"
        message:
          type: string
          example: "All canceled orders deleted successfully"

    BadRequestError:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Invalid request"

    NotFoundError:
      type: object
      properties:
        status:
          type: integer
          example: 404
        error:
          type: string
          example: "Order not found"

    ServerError:
      type: object
      properties:
        status:
          type: integer
          example: 500
        error:
          type: string
          example: "Internal server error"

    InvalidApiKeyError:
      type: object
      properties:
        error:
          type: string
          example: "Unauthorized"

    UnauthorizedError:
      type: object
      properties:
        error:
          type: string
          example: "Invalid username or password"

    ForbiddenError:
      type: object
      properties:
        status:
          type: integer
          example: 403
        error:
          type: string
          example: "Forbidden - buyer does not have access to this order"

    ConflictError:
      type: object
      properties:
        status:
          type: integer
          example: 409
        error:
          type: string
          example: "Order cannot be deleted unless status is CANCELED"

    XmlValidationResponse:
      type: object
      properties:
        valid:
          type: boolean
          example: true
        errors:
          type: array
          items:
            type: string
          example: []

    XmlValidationError:
      type: object
      properties:
        valid:
          type: boolean
          example: false
        errors:
          type: array
          items:
            type: string
          example:
            - "Missing required element: orderId"
            - "Invalid currency code"